refactor(cli): extract openInBrowser to shared utils
Deduplicate openInBrowser from create.ts and open.ts into utils.ts.
The shared version includes error handling (missing in create.ts) and
removes the unused spawn import from individual command files.
