+++ cli/src/commands/open.ts
@@ -2,9 +2,9 @@
  * zhe open <slug> — Open short URL in browser
  */
 
-import { spawn } from "node:child_process";
 import { defineCommand, pc } from "@nocoo/cli-base";
 import { EXIT_INVALID_ARGS } from "../api/client.js";
+import { openInBrowser } from "../utils.js";
 
 const BASE_URL = "https://zhe.to";
 
@@ -34,29 +34,3 @@ export const openCommand = defineCommand({
 		openInBrowser(url);
 	},
 });
-
-function openInBrowser(url: string): void {
-	const platform = process.platform;
-	let command: string;
-	let args: string[];
-
-	if (platform === "darwin") {
-		command = "open";
-		args = [url];
-	} else if (platform === "linux") {
-		command = "xdg-open";
-		args = [url];
-	} else if (platform === "win32") {
-		command = "cmd";
-		args = ["/c", "start", "", url];
-	} else {
-		console.log(pc.dim(`Unable to open browser on ${platform}. Visit: ${url}`));
-		return;
-	}
-
-	const child = spawn(command, args, { stdio: "ignore", detached: true });
-	child.unref();
-	child.on("error", () => {
-		console.log(pc.dim(`Failed to open browser. Visit: ${url}`));
-	});
-}
