+++ cli/src/utils.ts
@@ -2,6 +2,7 @@
  * Shared utilities for zhe CLI
  */
 
+import { spawn } from "node:child_process";
 import { pc } from "@nocoo/cli-base";
 import type { ApiClient } from "./api/client.js";
 import type { Folder, Link, Tag } from "./api/types.js";
@@ -263,6 +264,36 @@ export async function resolveFolderName(
 	}
 
 	return matches[0].id;
+}
+
+/**
+ * Open a URL in the default browser.
+ * Uses spawn() with argv to prevent shell injection.
+ */
+export function openInBrowser(url: string): void {
+	const platform = process.platform;
+	let command: string;
+	let args: string[];
+
+	if (platform === "darwin") {
+		command = "open";
+		args = [url];
+	} else if (platform === "linux") {
+		command = "xdg-open";
+		args = [url];
+	} else if (platform === "win32") {
+		command = "cmd";
+		args = ["/c", "start", "", url];
+	} else {
+		console.log(pc.dim(`Unable to open browser on ${platform}. Visit: ${url}`));
+		return;
+	}
+
+	const child = spawn(command, args, { stdio: "ignore", detached: true });
+	child.unref();
+	child.on("error", () => {
+		console.log(pc.dim(`Failed to open browser. Visit: ${url}`));
+	});
 }
 
 /**
